+++ pcs/app.py
@@ -64,7 +64,7 @@ def _non_root_run(argv_cmd):
         ["cluster", "disable", "..."],
         ["cluster", "enable", "..."],
         ["cluster", "node", "..."],
-        ["cluster", "pcsd-status", "..."],
+        ["cluster", "pcsd-status", "..."],  # TODO deprecated, remove command
         ["cluster", "start", "..."],
         ["cluster", "stop", "..."],
         ["cluster", "sync", "..."],
@@ -72,6 +72,7 @@ def _non_root_run(argv_cmd):
         ["host", "auth", "..."],
         ["host", "deauth", "..."],
         ["pcsd", "deauth", "..."],
+        ["pcsd", "status", "..."],
         ["pcsd", "sync-certificates"],
         ["quorum", "device", "status", "..."],
         ["quorum", "status", "..."],